User Creating

+++ src/app/(main)/admin/delete-application.tsx
@@ -1,18 +1,30 @@
 "use client";
 
 import { LoadingButton } from "@/components/loading-button";
-import { useTransition } from "react";
+import { useState, useTransition } from "react";
 import { toast } from "sonner";
 import { deleteApplication } from "./actions";
+import { Input } from "@/components/ui/input";
 
 export function DeleteApplication() {
   const [isPending, startTransition] = useTransition();
+  const [confirmText, setConfirmText] = useState("");
 
-  async function handleDeleteApplication() {
+  const isAllowed = confirmText === "DELETE";
+
+  function handleDeleteApplication() {
+    if (!isAllowed) {
+      toast.error('Type "DELETE" to confirm');
+      return;
+    }
+
     startTransition(async () => {
       try {
         await deleteApplication();
-        toast.success("Application deletion authorized successfully");
+        toast.success("Application deleted successfully");
+
+        // Full reload to reset the site
+        window.location.href = "/";
       } catch (error) {
         console.error(error);
         toast.error("Something went wrong");
@@ -31,13 +43,21 @@ export function DeleteApplication() {
             <p className="text-foreground/65 text-sm">
               This action will delete the entire application. This cannot be
               undone.
+              <br />
+              Type <b>DELETE</b> to confirm.
             </p>
           </div>
+          <Input
+            placeholder='Type "DELETE"'
+            value={confirmText}
+            onChange={(e) => setConfirmText(e.target.value)}
+          />
           <LoadingButton
             loading={isPending}
             onClick={handleDeleteApplication}
             variant="destructive"
             className="w-full"
+            disabled={!isAllowed}
           >
             Delete Application
           </LoadingButton>
